warning fixing frenzy mostly unused parameter, except GlCompositeHierarchyManager (initialization order) and TlpTools.cpp (unsigned int used instead of size_t, probable bug for x64 platforms)
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2697 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/TlpTools.cpp
@@ -277,7 +277,7 @@ std::string tlp::getMajor(const std::string& v) {
 
 //=========================================================
 std::string tlp::getMinor(const std::string& v) {
-  unsigned int pos = v.find('.');
+  size_t pos = v.find('.');
   if (pos == string::npos)
     return string("0");
   unsigned int rpos = v.rfind('.');
